+++ python/GafferTest/AnimationTest.py
@@ -403,9 +403,9 @@ class AnimationTest( GafferTest.TestCase ) :
 		k2 = Gaffer.Animation.Key( time = 1, value = 2 )
 		k3 = Gaffer.Animation.Key( time = 1, value = 3 )
 
-		curve.addKey( k1, False )
-		curve.addKey( k2, False )
-		curve.addKey( k3, False )
+		curve.addKey( k1, removeActiveClashing = False )
+		curve.addKey( k2, removeActiveClashing = False )
+		curve.addKey( k3, removeActiveClashing = False )
 
 		self.assertTrue( curve.getKey( 1 ).isSame( k3 ) )
 		self.assertTrue( k3.parent().isSame( curve ) )
@@ -519,9 +519,9 @@ class AnimationTest( GafferTest.TestCase ) :
 		k2 = Gaffer.Animation.Key( time = 1, value = 2 )
 		k3 = Gaffer.Animation.Key( time = 1, value = 3 )
 
-		curve.addKey( k1, False )
-		curve.addKey( k2, False )
-		curve.addKey( k3, False )
+		curve.addKey( k1, removeActiveClashing = False )
+		curve.addKey( k2, removeActiveClashing = False )
+		curve.addKey( k3, removeActiveClashing = False )
 
 		self.assertTrue( curve.getKey( 1 ).isSame( k3 ) )
 		self.assertTrue( k3.parent().isSame( curve ) )
@@ -607,9 +607,9 @@ class AnimationTest( GafferTest.TestCase ) :
 			k = Gaffer.Animation.Key( time = i, value = 1 )
 			k1 = Gaffer.Animation.Key( time = i, value = 2 )
 			k2 = Gaffer.Animation.Key( time = i, value = 3 )
-			curve.addKey( k2, False )
-			curve.addKey( k1, False )
-			curve.addKey( k, False )
+			curve.addKey( k2, removeActiveClashing = False )
+			curve.addKey( k1, removeActiveClashing = False )
+			curve.addKey( k, removeActiveClashing = False )
 			self.assertTrue( k.parent().isSame( curve ) )
 			self.assertTrue( k1.parent().isSame( curve ) )
 			self.assertTrue( k2.parent().isSame( curve ) )
@@ -807,7 +807,7 @@ class AnimationTest( GafferTest.TestCase ) :
 		self.assertFalse( k2.isActive() )
 
 		# then add second key back to same curve (first key becomes inactive)
-		curve.addKey( k2, False )
+		curve.addKey( k2, removeActiveClashing = False )
 
 		self.assertTrue( k1.parent().isSame( curve ) )
 		self.assertTrue( k2.parent().isSame( curve ) )
@@ -916,7 +916,7 @@ class AnimationTest( GafferTest.TestCase ) :
 		self.assertFalse( k2.isActive() )
 
 		# add third key with same time outside undo system so that first key becomes inactive
-		curve.addKey( k3, False )
+		curve.addKey( k3, removeActiveClashing = False )
 
 		self.assertTrue( k1.parent().isSame( curve ) )
 		self.assertTrue( k2.parent().isSame( curve ) )
@@ -981,7 +981,7 @@ class AnimationTest( GafferTest.TestCase ) :
 		self.assertEqual( k2.getTime(), 2 )
 
 		# add third key with same time outside undo system so that first key becomes inactive
-		curve.addKey( k3, False )
+		curve.addKey( k3, removeActiveClashing = False )
 
 		self.assertTrue( k1.parent().isSame( curve ) )
 		self.assertTrue( k2.parent().isSame( curve ) )
@@ -1469,10 +1469,10 @@ class AnimationTest( GafferTest.TestCase ) :
 		k2 = Gaffer.Animation.Key( time = 2 )
 		k3 = Gaffer.Animation.Key( time = 2 )
 
-		curve.addKey( k0, False )
-		curve.addKey( k1, False )
-		curve.addKey( k3, False )
-		curve.addKey( k2, False )
+		curve.addKey( k0, removeActiveClashing = False )
+		curve.addKey( k1, removeActiveClashing = False )
+		curve.addKey( k3, removeActiveClashing = False )
+		curve.addKey( k2, removeActiveClashing = False )
 
 		self.assertTrue( curve.getKey( 1 ).isSame( k1 ) )
 		self.assertTrue( curve.getKey( 2 ).isSame( k2 ) )