+++ src/GafferSceneUI/CameraTool.cpp
@@ -89,7 +89,7 @@ void setValueOrAddKey( Gaffer::FloatPlug *plug, float time, float value )
 	if( Animation::isAnimated( plug ) )
 	{
 		Animation::CurvePlug *curve = Animation::acquire( plug );
-		curve->addKey( new Animation::Key( time, value ) );
+		curve->addKey( new Animation::Key( time, value ), /* inheritInterpolation = */ true );
 	}
 	else
 	{