+++ Changes.md
@@ -57,6 +57,7 @@ API
 - Animation :
   - `curve.addKey()` and `key.setTime()` now return active clashing key
   - `curve.addKey()` now has optional remove parameter that defaults to true, if false active clashing key is not removed and becomes inactive.
+  - `curve.addKey()` now has optional inherit parameter that defaults to false, if true then key's interpolation will be set to match span interpolation.
   - `key.setTime()` will not remove active clashing key, instead it becomes become inactive. inactive key at old time becomes active.
   - Added `curve.removeInactiveKeys()` function, removes all the inactive keys parented to a curve.
   - Added `key.isActive()` function, returns true if key is parented to curve and is the active key at its current time.